+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * feat: (library) adds MessageBag to store messages to log
 * feat: (library) adds AWSValidator to validate common AWS resource names
 * feat: (import-csv) Validates most of the fields of the CSV file
+* feat: (import-csv) Adds CSV Intak data class
 * refactor: (import-csv) changes script to use the cloud_migration_requester
 * refactor(BREAKING): changes input CSV headers fields to be all snake_case
 * doc: link to technical documentation